# REFACTOR - utils 파일을 하나로 합침
## 수정사항
- my_tools 제거하고 my_utils로 합침
- tx_generator 타입 안맞는것 수정
